Extract library instrumentation for lettuce 5.1
