+++ instrumentation/lettuce/lettuce-5.1/javaagent/lettuce-5.1-javaagent.gradle
@@ -12,7 +12,9 @@ muzzle {
 dependencies {
   library group: 'io.lettuce', name: 'lettuce-core', version: '5.1.0.RELEASE'
 
-  implementation project(':instrumentation:lettuce:lettuce-common:javaagent')
+  implementation project(':instrumentation:lettuce:lettuce-5.1:library')
+
+  testImplementation project(':instrumentation:lettuce:lettuce-5.1:testing')
 
   testImplementation group: 'com.github.kstyrc', name: 'embedded-redis', version: '0.6'
   // Only 5.2+ will have command arguments in the db.statement tag.
