+++ instrumentation/lettuce/lettuce-5.1/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/lettuce/v5_1/LettuceInstrumentationModule.java
@@ -59,7 +59,7 @@ public class LettuceInstrumentationModule extends InstrumentationModule {
 
     @Advice.OnMethodExit(suppress = Throwable.class)
     public static void methodEnter(@Advice.Return DefaultClientResources.Builder builder) {
-      builder.tracing(OpenTelemetryTracing.INSTANCE);
+      builder.tracing(TracingHolder.TRACING);
     }
   }
 }
