+++ instrumentation/lettuce/lettuce-5.1/library/src/test/groovy/io/opentelemetry/instrumentation/lettuce/v5_1/LettuceAsyncSyncClientTest.groovy
@@ -1,0 +1,21 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.lettuce.v5_1
+
+import io.lettuce.core.RedisClient
+import io.lettuce.core.resource.ClientResources
+import io.opentelemetry.instrumentation.test.LibraryTestTrait
+
+class LettuceAsyncSyncClientTest extends AbstractLettuceAsyncClientTest implements LibraryTestTrait {
+  @Override
+  RedisClient createClient(String uri) {
+    return RedisClient.create(
+      ClientResources.builder()
+        .tracing(LettuceTracing.create(getOpenTelemetry()).newTracing())
+        .build(),
+      uri)
+  }
+}
