+++ settings.gradle
@@ -157,10 +157,12 @@ include ':instrumentation:khttp-0.1:javaagent'
 include ':instrumentation:kotlinx-coroutines:javaagent'
 include ':instrumentation:kubernetes-client-7.0:javaagent'
 include ':instrumentation:kubernetes-client-7.0:javaagent-unittests'
-include ':instrumentation:lettuce:lettuce-common:javaagent'
+include ':instrumentation:lettuce:lettuce-common:library'
 include ':instrumentation:lettuce:lettuce-4.0:javaagent'
 include ':instrumentation:lettuce:lettuce-5.0:javaagent'
 include ':instrumentation:lettuce:lettuce-5.1:javaagent'
+include ':instrumentation:lettuce:lettuce-5.1:library'
+include ':instrumentation:lettuce:lettuce-5.1:testing'
 include ':instrumentation:liberty:compile-stub'
 include ':instrumentation:liberty:liberty:javaagent'
 include ':instrumentation:liberty:liberty-dispatcher:javaagent'
